pyproject.toml
[project]
name = "cck-agents"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "claudette>=0.1.4",
    "colbert-ai>=0.2.21",
    "docling>=2.15.1",
    "fastcore>=1.8.0",
    "ipykernel>=6.29.5",
    "ipywidgets>=8.1.5",
    "jupyter>=1.1.1",
    "langchain-text-splitters>=0.3.8",
    "matplotlib>=3.10.1",
    "monsterui>=1.0.18",
    "multidict==6.4.0",
    "nbdev>=2.3.37",
    "numpy==2.2.4",
    "python-dotenv>=1.1.0",
    "python-fasthtml>=0.12.12",
    "ragatouille>=0.0.9",
    "sentence-transformers>=4.0.2",
    "torch>=2.6.0",
    "torchaudio>=2.6.0",
    "torchvision>=0.21.0",
    "transformers>=4.51.1",
    "umap-learn>=0.5.7",
]

[tool.uv.sources]
claudette = { git = "https://github.com/KronResearch/claudette.git" }
